Add reactor dependency to domain layer
BookService interface is intended to be reactive, so its methods can be implemented accordingly in the data provider layer.
A first approach configuring reactive-mongo is taken, but it still needs work to make it happen (currently the app context starts-up well, but mongodb throws connection refused)
Unit tests in domain are adapted to reactive.